% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brocks.R
\name{ac_se}
\alias{ac_se}
\title{Agresti-Coull Intervals}
\usage{
ac_se(logical_var, wt = 2)
}
\arguments{
\item{logical_var}{A \code{\link{logical}} \code{\link{vector}}}

\item{wt}{The number of successes and failures to add to the sample before
construction of a Wald interval}
}
\value{
\code{\link{numeric}}. An estimate of the sample's standard error.
}
\description{
Agresti-Coull (1998) intervals are a great way to get a quick and
non-terrible estimate of a proportion. They work by using a 'Wald' interval,
after the addition of 2 successes and 2 failures to the sample (other numbers
can be specified, via the \code{wt} argument).
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}
\references{
{Ag, F., & Dewaele, J.M. (2002). Variation in the contextuality of language: An empirical measure. Context in Context, Special issue of Foundations of Science, 7 (3), 293-340.}
}

