% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{add_line}
\alias{add_line}
\title{Taken from https://github.com/hadley/httr/blob/1fc659856602f60ff75eb01903513244e3491ec2/R/oauth-cache.R#L52}
\usage{
add_line(path, line, quiet = FALSE)
}
\description{
Taken from https://github.com/hadley/httr/blob/1fc659856602f60ff75eb01903513244e3491ec2/R/oauth-cache.R#L52
}
\keyword{internal}

