% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blog_stuff.R
\name{blog_opts}
\alias{blog_opts}
\title{Set some knitr chunk options which may work well for blogging}
\usage{
blog_opts(...)
}
\arguments{
\item{...}{passed to knitr::\code{\link[knitr]{opts_chunk}}$set}
}
\description{
A small wrapper around knitr's \code{\link[knitr]{opts_chunk}}$set, with some
defaults which I've found work well for blog posts. All messages from R are
surpressed, and the quality of the plots is increased to 6" X 6" 300 dpi
\code{\link{png}}s.
}

