% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blog_stuff.R
\name{blog_push}
\alias{blog_push}
\title{Push a blog post live (possibly)}
\usage{
blog_push(command = "bash _deploy.sh staging")
}
\arguments{
\item{command}{Something which will be executed by \code{\link{system}}}
}
\value{
Used for its side effects.
}
\description{
I use this function to push blog posts live, as on my machine I have a bash
alias \code{blog_push} which does this. This is an incredibyly lazy wrapper
for \code{blog_gen();system(command)}, where \code{command} is by default
how I upload changes to my website.
}

