% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blog_stuff.R
\name{blog_run}
\alias{blog_run}
\title{Serve or Compile a Jekyll Blog}
\usage{
blog_run(input = c(".", list.dirs("_source")), output = c(".", rep("_posts",
  length(list.dirs("_source")) - 1)), ...)
}
\arguments{
\item{input}{passed to servr::\code{\link{jekyll}}}

\item{output}{passed to servr::\code{\link{jekyll}}}

\item{...}{passed to servr::\code{\link{jekyll}}}
}
\description{
Serve or Compile a Jekyll Blog. A small wrapper around
servr::\code{\link{jekyll}}, which be default looks for subdirectories witin
the main source directory.
}

