% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brocks.R
\name{char_cols}
\alias{char_cols}
\title{Convert all columns in a data.frame to character}
\usage{
char_cols(x, all = FALSE)
}
\arguments{
\item{x}{A \code{data.frame}}

\item{all}{Should all the columns (not just those of \code{factor}s) be
  converted to \code{character}?}
}
\description{
\bold{\code{R}}'s tendency to convert strings to \code{factor}s is well
meaning, but occasionally annoying when munging data (e.g. when using
\code{\link{rbind}}, or \code{dplyr}'s \code{\link[dplyr]{bind_rows}}).
\code{char_cols} turns all \code{factor} variables in a
\code{\link{data.frame}} into \code{\link{character}} variables. Optionally,
it will turn all variables to \code{character} if the parameter \code{all} is
set to \code{TRUE}.
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}

