% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brocks.R
\name{consolidate_values}
\alias{consolidate_values}
\title{Attempt to Rationalise a Factor Variable (using simple rules)}
\usage{
consolidate_values(x, case = "lower",
  na_regex = "no info*|don't know|<na>|#na|n/a")
}
\arguments{
\item{x}{The \code{\link{factor}} variable which you'd like to consolidate
the levels of}

\item{case}{The case which the levels of \code{x} will be converted to. Can
be one of \code{"lower"}, \code{"upper"}.}

\item{na_regex}{A regular expression which will be used to set values of
  \code{x} to \code{NA}.}
}
\value{
\code{\link{factor}}, hopefully with consolidated levels
}
\description{
{
  \code{consolidate_values} aims to do some of the work of the grunt work of
  rationalising factor levels for you. Using some very simple rules, it will
  attempt to turn a vector containing what appear to be gramatical variations
  on common patterns, into one with these rationalised to a smaller set of
  values.

  \itemize{
    \item All strings are converted to the same case (lower, by default)
    \item Spaces around slashes, before commas, or after common currency
      symbols are removed
    \item All underscores and full stops are replaced with spaces
    \item All continuous whitespace is converted to a single space
    \item All trailing and leading whitespace is removed
    \item Blank values ("") are converted to \code{NA}
    \item Any strings which match \code{na_regex} are converted to \code{NA}.
  }
}
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}

