% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/munge.R
\name{date_to_age}
\alias{date_to_age}
\title{Convert Dates to Ages}
\usage{
date_to_age(dob, refdate = Sys.time())
}
\arguments{
\item{dob}{'Date of Birth'}

\item{refdate}{Reference date to compare \code{x} with. Defaults to the
current date}
}
\value{
Numeric, 'whole number' ages
}
\description{
Convert dates to ages.
}
\examples{
# The argument can take a single age
date_to_age("1952-04-29")

# Or a whole vector!
data(test_data)
test_data$age <- date_to_age(test_data$dob)

# Historical ages, too. How old were these people yesterday?

test_data$age <- date_to_age(test_data$dob, Sys.Date() - 1)
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}

