% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extract_package_deps}
\alias{extract_package_deps}
\title{Extract package dependencies from an R script}
\usage{
extract_package_deps(file)
}
\arguments{
\item{file}{A file containing R code to parse}
}
\value{
A character vector containing the names of packages used in
  \code{file}
}
\description{
Extract package dependencies from an R script
}

