% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{geom_watermark}
\alias{geom_watermark}
\title{Add a watermark to a plot}
\usage{
geom_watermark(png_path = Sys.getenv("WATERMARK_PNG"))
}
\arguments{
\item{png_path}{A path to a valid png file to use for the watermark}
}
\description{
Add a watermark to a plot
}

