% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{gg_colour_hue}
\alias{gg_colour_hue}
\title{ggplot2 categorical colour hues}
\usage{
gg_colour_hue(n)
}
\arguments{
\item{n}{The number of colours to return}
}
\description{
Generate the colours which \code{ggplot2} would, given a certain number of
categories.
}
\examples{
barplot(1:10, col = gg_colour_hue(10))
}
\author{
John Colby (taken from: \url{https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette})
}

