% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{gg_footnote}
\alias{gg_footnote}
\title{Add a footnote to a ggplot chart}
\usage{
gg_footnote(p, text = "This is a footnote", detect_font = TRUE,
  wrap = TRUE, newlines = TRUE, fontface = "italic", fontsize = 8,
  col = "grey50", fontfamily = NULL, ...)
}
\arguments{
\item{p}{A ggplot2 plot object}

\item{text}{The text used for the footnote}

\item{detect_font}{for ggplot2 objects, should the function attempt to
extract the default font used?}

\item{wrap}{Should line breaks be added to the text to (attempt) to prevent
it overflowing?}

\item{newlines}{Should a new line be appended to the start and end of the
string, for spacing?}

\item{fontface}{Passed to \code{\link{gpar}}}

\item{fontsize}{Passed to \code{\link{gpar}}}

\item{col}{Passed to \code{\link{gpar}}}

\item{fontfamily}{Passed to \code{\link{gpar}}}

\item{...}{Passed to \code{\link{gpar}}}
}
\description{
Currently, the text wrapping is based on the plot's default font size (which
is usually 12pts), meaning that using something smaller (as is the default in
this function), is likely to result in text which is slightly more narrow
than the available horizontal space. I don't mind it. Someone smarter than I
might be able to fix it with splitTextGrob!
}
\examples{
\dontrun{
  # Let's make a simple plot and add a footnote to it.

  # Some footnote text, taken from the 'fortunes' package
  foot_note <- "R is Open Source and so you can modify it to emulate the bugs
  in other software: that is not one of the aims of its developers so please
  don't expect us to do so for you.\\n  -- Brian D. Ripley (answering a request
  for a change to Excel-like non IEC 60559 standard conform rounding)
  \\n     R-help (March 2009)"

  # A ggplot2 chart
  library(ggplot2)
  d <- data.frame(mtcars)
  p <- ggplot(d, aes(cyl, mpg)) +
  geom_point() +
  theme(text = element_text(family = "Garamond"))

  # Let's add our footnote!
  g <- gg_footnote(p, foot_note)
  g

  # Note that the text is horizontally truncated. It will take up the full
  # width if we use the default ggplot2 font size of 12pts
  g <- gg_footnote(p, foot_note, fontsize = 12)
  g
}
}

