% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{guess_db_connection}
\alias{guess_db_connection}
\title{If there's only one database connection, use it}
\usage{
guess_db_connection(db_con_classes = "PostgreSQLConnection")
}
\description{
If there's only one database connection, use it
}
\keyword{internal}

