% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blog_stuff.R
\name{htmlwidgets_deps}
\alias{htmlwidgets_deps}
\title{Configure htmlwidgets dependencies for a knitr-jekyll blog}
\usage{
htmlwidgets_deps(a, knit_meta = knitr::knit_meta(),
  lib_dir = "htmlwidgets_deps", includes_dir = "_includes/htmlwidgets/",
  always = FALSE)
}
\arguments{
\item{a}{The file path for the input file being knit}

\item{knit_meta}{The dependencies object.}

\item{lib_dir}{The directory where the htmlwidgets dependency source code can be found (e.g. JavaScript and CSS files)}

\item{includes_dir}{The directory to add the HTML file to}

\item{always}{Should dependency files always be produced, even if htmlwidgets are not being used?}
}
\value{
Used for it's side effects.
}
\description{
This is really just added to build.R.
}

