% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{install_deps}
\alias{install_deps}
\title{Recursively Parse R Files in a Directory, and Install Packages Used}
\usage{
install_deps(dir = getwd(), file_pattern = "\\\\.R$|\\\\.Rmd$", ...)
}
\arguments{
\item{dir}{The directory to search for R files to parse}

\item{file_pattern}{A regular expression used to determine whether a file
should be parsed or not. The default will parse only \code{.R} and
\code{.Rmd} files}

\item{...}{Passed to \code{\link{install.packages}}}
}
\value{
Used for it's side effects (the installation of packages)
}
\description{
\code{install_deps} (recursively) finds R code files in a directory, and uses
regular expressions to find code that looks like it refers to an R package
(via \code{\link{extract_package_deps}}). It then extracts the names of all
of these packages, checks that they're not already installed, and that they
are on CRAN, and then installs them (via \code{\link{install.packages}}).
}

