% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brocks.R
\name{kill_cols}
\alias{kill_cols}
\alias{killcols}
\title{Kill off non-informative columns in a data.frame}
\usage{
kill_cols(x, empty_strings = TRUE)

killcols(x, empty_strings = TRUE)
}
\arguments{
\item{x}{A \code{data.frame}}

\item{empty_strings}{\code{logical}. Should empty strings ("") be converted
to \code{NA}?}
}
\value{
A \code{data.frame}, less non-varying variables.
}
\description{
\code{kill_cols} accepts a \code{\link{data.frame}} or \code{\link{matrix}}, and returns another, containg only the columns with more than one unique non-missing value. By default, empty strings ("") are converted to \code{NA}, and not considered informative.
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}

