% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brocks.R
\name{lz}
\alias{lz}
\title{Turn numbers to strings, with leading zeros}
\usage{
lz(x, n = nchar(max(as.numeric(x))))
}
\arguments{
\item{x}{A \code{\link{numeric}} \code{\link{vector}} (non numeric values
will be coerced)}

\item{n}{The character length of the returned strings. Defaults to the length
of the maximum value of x}
}
\value{
A \code{\link{character}} \code{\link{vector}} of those numbers, with
  leading zeros
}
\description{
Sometimes useful for creating composite primary keys, which order nicely.
}
\examples{
lz(c(1, 10, 100))
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}

