% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{misc_br_num_formats}
\alias{fmt_nps}
\alias{fmt_nps_pm}
\alias{fmt_pc}
\alias{fmt_pc_pm}
\alias{fmt_pm}
\alias{misc_br_num_formats}
\alias{unsci}
\alias{unsci_dollars}
\title{Miscellaneous Number Formatting Functions}
\usage{
fmt_pm(x, ...)

fmt_pc(x, type = c("round", "greater", "less"), digits = 0)

fmt_pc_pm(x, ...)

fmt_nps_pm(x, ...)

fmt_nps(x, ...)

unsci(x, digits = 1, currency = FALSE, symbol = "$", pad = TRUE)

unsci_dollars(x, ...)
}
\arguments{
\item{x}{\code{\link{numeric}} data to format}

\item{...}{Passed to \code{\link{round}}}

\item{type}{The truncation function for the number, in the context of the
text in which its likely to be formatted. One of \code{round} (the
\code{round} function is used), \code{greater} (the \code{floor} function
is used) or \code{less} (the \code{ceiling} function is used).}

\item{digits}{Parameter passed to \code{\link{round}}}

\item{currency}{Should numbers be prefixed with \code{symbol}?}

\item{symbol}{if \code{currency = TRUE}, a string to prefix numbers with}

\item{pad}{Should the resulting strings be prefix-padded with spaces to make
all strings in the character vector a uniform width?}
}
\value{
\code{\link{character}}.
}
\description{
{
  Sometimes (for example when illustrating differences), it can be useful for
  positive numbers to be prefixed by a + sign, just as negative numbers are
  with a - sign. The following are a few (very simple) wrapper functions
  which do this.

  \describe{
    \item{\bold{\code{fmt_pm}}}{ Is a wrapper for \code{\link{round}},
      which also \code{\link{paste}}s a + sign before positive numbers
    }
    \item{\bold{\code{fmt_pc}}}{ A simple formatting function for
      percentages. Defaults to 0 decimal places
    }
    \item{\bold{\code{fmt_pc_pm}}}{ As above, but with a + prefix for
      positive numbers
    }
    \item{\bold{\code{format_nps}}}{ A very simple formatter for the Net
      Promoter Score
    }
    \item{\bold{\code{fmt_nps_pm}}}{ As above, but without the percentage
      sign
    }
    \item{\bold{\code{unsci}}}{ Unscientific notation: Short colloquial
      number formatting. For example, 1e+04 becomes "100k", 1.454e+09 becomes
      "1.5B", etc.
    }
    \item{\bold{\code{unsci_dollars}}}{ A convenience function for the above,
      with \code{currency = TRUE} as the default.
    }
  }
}
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}

