% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blog_stuff.R
\name{new_post}
\alias{new_post}
\title{File Structure for a Jekyll Blog Post}
\usage{
new_post(title = "new post", serve = TRUE, dir = "_source",
  subdir = TRUE, skeleton_file = ".skeleton_post")
}
\arguments{
\item{title}{The title of the blog post}

\item{serve}{Should \code{\link{blog_serve}} be run once the files have
been set-up? Defatuls to \code{TRUE}.}

\item{dir}{The directory the post (or subdirectory) should reside in}

\item{subdir}{Should the post live in a subdirectory? Defaults to \code{TRUE}}

\item{skeleton_file}{The filepath of a skeleton blog post which will be used
as the basis for the basis for the newly created file}
}
\description{
File Structure for a Jekyll Blog Post
}
\details{
{
  \code{new_post} will create a .R file, and a .Rmd file (by default in a
  subdirectory), with names created by running \code{title} through
  \code{\link{filenamise}}. The .R file will contain a short note mentioning
  that it accompanies the .Rmd file, which will contain the same text as the
  file supplied by \code{skeleton_post} paramter. Both files will be opened
  using \code{\link{sys_open}}.
}
}

