% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_gs}
\alias{read_gs}
\title{A wrapper for the googlesheets package: A lazy way to read a googlesheet}
\usage{
read_gs(key = NULL, title = NULL, url = NULL)
}
\arguments{
\item{key}{Passed to \code{\link{gs_key}}}

\item{title}{Passed to \code{\link{gs_title}}}

\item{url}{Passed to \code{\link{gs_url}}}
}
\value{
The results of \code{\link{gs_read}}
}
\description{
This auths (via \code{\link{gs_auth}}), finds a sheet, and reads it (via
\code{\link{gs_read}}). It assumes that you've already set-up your computing
evironment to use the the \code{\link{googlesheets}} pacakge; see it's
documentation for more details.
}

