% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_txt}
\alias{read_txt}
\title{Read a text file}
\usage{
read_txt(file, print = FALSE)
}
\arguments{
\item{file}{A filename to read-in}

\item{print}{Should the contents of the file be printed to the console with
\code{\link{cat}}? (If so, results will be returned invisibly.)}
}
\value{
The contents of the file, as a character vector of length one.
}
\description{
A simple wrapper function around \code{\link{readLines}}, which combines each
line into a single string, via \code{paste0(x, collapse = "\n")}.
}

