% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rep_char}
\alias{rep_char}
\title{Repeat a character a variable number of times}
\usage{
rep_char(x = " ", times)
}
\arguments{
\item{x}{A value to repeat. Will be coerced to \code{\link{character}}.}

\item{times}{A \code{\link{numeric}} \code{\link{vector}}; the number of
times that \code{x} should be repeated.}
}
\value{
A \code{\link{character}} \code{\link{vector}} of x repated various
  times
}
\description{
Effectively a version of \code{\link{rep}}, where only once value can be
repeated (by default, a space; " "), but it can be repeated a variable number
of times. Useful for creating even spacing for print and summary methods.
}
\examples{

# Strings repeating 'a' a variable number of times!
rep_char("a", 1:5)

# Slightly more useful. Some text strings which we'd like to present:
desc <- c("\\n", "first : 1st\\n", "second : 2nd\\n", "third : 3rd\\n",
"fourth : 4th\\n", "umpteenth : ...\\n")

# However, the varying lengths make them look a little awkward
cat(desc)

# We can use rep_char to add extra spaces to the strings which are shorter
# than the longest
desc_spaced <- paste0(rep_char(times = max(nchar(desc)) - nchar(desc)), desc)

# Much better
cat(desc_spaced)

}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}

