% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{rev_levs}
\alias{rev_levs}
\title{Reverse Factor Levels}
\usage{
rev_levs(x)
}
\arguments{
\item{x}{A \code{\link{factor}} \code{\link{vector}}. If not, it will be
coerced.}
}
\value{
A A \code{\link{factor}} \code{\link{vector}}, with the ordering of
  the levels reversed
}
\description{
Reverse the ordering of levels in a factor variable. Sometimes useful when
plotting data.
}
\examples{
levels(rev_levs(1:10))
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}

