% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/munge.R
\name{scale_strip}
\alias{scale_strip}
\title{Strip Scale Labels from Likert Data}
\usage{
scale_strip(x)
}
\arguments{
\item{x}{a \code{\link{vector}}, \code{\link{matrix}}, or
\code{\link{data.frame}} containg \code{\link{character}} data}
}
\value{
\code{\link{numeric}} values stripped from the \code{\link{character}}
  values in \code{x}.
}
\description{
Survey systems commonly export data files with values measured by Likert
scales coded as \code{\link{character}} as opposed to \code{\link{numeric}},
(e.g. \code{"10 - Very Likely"} as opposed to \code{10}). The
\code{scale_strip} function pulls the first sequence of digits from a string
and returns it as numeric.
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}

