% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{smx_watermark}
\alias{smx_watermark}
\title{Add a watermark to a plot}
\usage{
smx_watermark(png_path = "~/projects/smx_logos/greyscale_top.png")
}
\arguments{
\item{png_path}{A path to a valid png file to use for the watermark}
}
\description{
Add a watermark to a plot
}

