% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blog_stuff.R
\name{sys_open}
\alias{sys_open}
\title{Open a file in the Operating System}
\usage{
sys_open(f)
}
\arguments{
\item{f}{The path of the file to open}
}
\value{
Nothing. Used for it's side effect.
}
\description{
Open a file in the operating system, using it's default program.
\code{sys_open} should work on Ubuntu (and other Linux variants), OSX and
Windows.
}
\author{
Based very heavily on the function \code{openFileInOS} from the
  package \code{pander} (v0.5.2), written by Gergely Daroczi
  (\email{daroczig@rapporter.net}), itself based on the \code{convert}
  function in the package \code{ascii}, written by David Hajage
  (\email{dhajage@gmail.com}).
}

