% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{to_sql_vector}
\alias{to_sql_vector}
\title{Turn and R vector into a SQL vector}
\usage{
to_sql_vector(x)
}
\arguments{
\item{x}{A vector}
}
\value{
\code{character}
}
\description{
Turn and R vector into a SQL vector
}
\examples{
to_sql_vector(letters[1:10])
}

