% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blog_stuff.R
\name{blog_push}
\alias{blog_push}
\title{Push a blog post live (possibly)}
\usage{
blog_push(command = "blog_push")
}
\arguments{
\item{command}{Something which will be executed by \code{\link{system}}}
}
\value{
Used for its side effects.
}
\description{
I use this function to push blog posts live, as on my machine I have a bash
alias \code{blog_push} which does this. This is an incredibyly lazy wrapper
for \code{blog_gen();system(command)}, where \code{command} is by default the
bash alias I use.
}
\details{
{
  If it's of any interest, my \code{push_blog} bash alias is:

  \code{'aws s3 sync /home/br/projects/brendanrocks.com/_site \
         s3://brendanrocks.com --exclude "cache/*|README.md" --delete'}
}
}

