% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blog_stuff.R
\name{blog_serve}
\alias{blog_gen}
\alias{blog_serve}
\alias{site_gen}
\alias{site_serve}
\title{Serve or Compile a Jekyll Blog}
\usage{
blog_serve(input = c(".", list.dirs("_source")), output = c(".",
  rep("_posts", length(list.dirs("_source")))), ...)

blog_gen(input = c(".", list.dirs("_source")), output = c(".", rep("_posts",
  length(list.dirs("_source")))), ...)

site_gen(input = c(".", list.dirs("_dashboards"), list.dirs("_source")),
  output = c(".", rep(".", length(list.dirs("_dashboards"))), rep("_posts",
  length(list.dirs("_source")))), ...)

site_serve(input = c(".", list.dirs("_dashboards"), list.dirs("_source")),
  output = c(".", rep(".", length(list.dirs("_dashboards"))), rep("_posts",
  length(list.dirs("_source")))), ...)
}
\arguments{
\item{input}{passed to servr::\code{\link{jekyll}}}

\item{output}{passed to servr::\code{\link{jekyll}}}

\item{...}{passed to servr::\code{\link{jekyll}}}
}
\description{
Serve or Compile a Jekyll Blog. A small wrapper around
servr::\code{\link{jekyll}}, which by default also looks for subdirectories
witin the main source directory.
}

