% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{get_query}
\alias{get_query}
\title{Retreive a database query from a SQL statement in a text file}
\usage{
get_query(sql_file, con = guess_db_connection(), local_cache = TRUE,
  cache_dir = "./.sql_cache", gitignore = TRUE)
}
\arguments{
\item{sql_file}{A path to a file to SQL}

\item{con}{A database connection. If unspecified, and there is a single
database connection in the global environment, this will be used}

\item{local_cache}{Should results be cached locally as \code{.RData} files
via \code{\link{saveRDS}}?}

\item{cache_dir}{The directory in which to cache files}

\item{gitignore}{Should the cache directory be \code{.gitignored}?}
}
\value{
The results of passing the SQL statement to \code{\link{dbGetQuery}}
}
\description{
A convenience function which wraps \code{\link{dbGetQuery}}, by default
providing some local caching
}

