% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{misc_br_num_formats}
\alias{format_nps}
\alias{misc_br_num_formats}
\alias{plus_minus}
\alias{plus_minus_nps}
\alias{plus_minus_percent}
\title{Miscellaneous Number Formatting Functions}
\usage{
plus_minus(x, ...)

plus_minus_percent(x, ...)

plus_minus_nps(x, ...)

format_nps(x, ...)
}
\arguments{
\item{x}{\code{\link{numeric}} data to format}

\item{...}{Passed to \code{\link{round}}}
}
\value{
\code{\link{character}}.
}
\description{
{
  Sometimes (for example when illustrating differences), it can be useful for
  positive numbers to be prefixed by a + sign, just as negative numbers are
  with a - sign. The following are a few (very simple) wrapper functions
  which do this.

  \describe{
    \item{\bold{\code{plus_minus}}}{ Is a wrapper for \code{\link{round}},
      which also \code{\link{paste}}s a + sign before positive numbers
    }
    \item{\bold{\code{plus_minus_percent}}}{ Does the same as above, but
      multiplies by 100 first, and adds a % sign to the end of the number
    }
    \item{\bold{\code{plus_minus_nps}}}{ Does the same as the above, but
      without the percentage sign
    }
    \item{\bold{\code{format_nps}}}{ Does the same as the above, but without
      the + suffix for positive numbers
    }
  }
}
}
\examples{

plus_minus(-3:3)
plus_minus_percent(runif(6) - .5)
plus_minus_nps(runif(6) - .5)
format_nps(runif(6) - .5)

}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}

