% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brocks.R
\name{refactor}
\alias{refactor}
\title{Change Factor Levels Using List Lookups}
\usage{
refactor(x, new_values, throw_error = FALSE)
}
\arguments{
\item{x}{A \code{\link{factor}} variable}

\item{new_values}{A list}

\item{throw_error}{If there are values in \code{x} which are not explicitly
  mapped in \code{new_values}, should the function throw an error?}
}
\value{
\code{factor}
}
\description{
Change Factor Levels Using List Lookups
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}

