% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brocks.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{An idealised test data set, for demonstrating some of the functions}
\description{
An idealised test data set, for demonstrating some of the functions in the
package
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}
\keyword{data}

